[#8] Alert message for delete category created

+++ Expired/Views/CategoriesView.swift
@@ -10,9 +10,10 @@ import SwiftUI
 struct CategoriesView: View {
     @Environment(\.managedObjectContext) private var viewContext
     @EnvironmentObject var productStore: ProductStore
-
-    @State private var showingFormAlert: Bool = false
-    @State private var category: String = ""
+    @State private var showingNotDeleteAlert: Bool = false
+//    @State private var category: String = ""
+    @State private var showingDeleteAlert = false
+    @State private var deleteIndexSet: IndexSet?
     
     var emptyPlaceholderText: String = "No category found\nPress + to add your first category!"
 
@@ -20,9 +21,15 @@ struct CategoriesView: View {
         NavigationView {
             List {
                 ForEach(productStore.categories) { category in
-                    Text(category.title ?? "")
+                    VStack(alignment: .leading) {
+                        Text(category.title ?? "")
+                            .font(.headline)
+                        Text(category.products?.count ?? 0 > 0 ? "\(category.products!.count) products" : "No product")
+                            .font(.subheadline)
+                            .foregroundColor(.gray)
+                    }
                 }
-                .onDelete(perform: deleteCategory)
+                .onDelete(perform: showDeleteAlert)
             }
             .overlay(Group {
                 if productStore.categories.isEmpty {
@@ -31,6 +38,23 @@ struct CategoriesView: View {
                         .multilineTextAlignment(.center)
                         .padding()
                 }
+            })
+            .alert("Are you sure you want to delete this category?", isPresented: $showingDeleteAlert) {
+                Button("Maybe Later", role: .cancel) {
+                    deleteIndexSet = nil
+                }
+                Button("Yes", role: .destructive) {
+                    if let indexSet = deleteIndexSet {
+                        deleteCategory(indexSet: indexSet)
+                    }
+                    deleteIndexSet = nil
+                }
+            }
+            .alert("Unable To Delete", isPresented: $showingNotDeleteAlert , actions: {
+                Button("Ok", role: .cancel) {
+                }
+            }, message: {
+                Text("Some products are linked to this category")
             })
             .navigationTitle("Categories")
             .toolbar {
@@ -45,13 +69,21 @@ struct CategoriesView: View {
         }
     }
     private func deleteCategory(indexSet: IndexSet){
-            withAnimation {
-                indexSet.map{productStore.categories[$0]}.forEach{ category in
+        withAnimation {
+            indexSet.map{productStore.categories[$0]}.forEach{ category in
+                if(category.products?.count == 0){
                     viewContext.delete(category)
+                }else{
+                   showingNotDeleteAlert =  true
                 }
                 productStore.save(viewContext)
             }
         }
+    }
+    private func showDeleteAlert(indexSet: IndexSet) {
+        deleteIndexSet = indexSet
+        showingDeleteAlert = true
+    }
 }
 
 struct CategoriesView_Previews: PreviewProvider {
